// src/modules/checkout/usecase/place-order/place-order.dto.ts
import Product from "../../domain/product.entity"

export interface PlaceOrderInputDto {
    clientId: string
    products: {
        productId: string
    }[]
}

export interface PlaceOrderOutputDto {
    id: string
    invoiceId: string
    status: string
    total: number
    products: {
        productId: string
    }[]
}